+++ src/Serializers/Igbinary.php
@@ -6,16 +6,16 @@ use Rubix\Server\Message;
 use RuntimeException;
 
 /**
- * Binary
+ * Igbinary
  *
- * Converts persistable object to and from a binary encoding. Binary format is
+ * Converts persistable object to and from a binary encoding. Igbinary format is
  * smaller and typically faster than plain text serializers.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class Binary implements Serializer
+class Igbinary implements Serializer
 {
     /**
      * @throws \RuntimeException